+++ payments-core/src/test/java/com/stripe/android/StripeKtxTest.kt
@@ -2,14 +2,14 @@ package com.stripe.android
 
 import android.content.Intent
 import com.google.common.truth.Truth.assertThat
+import com.stripe.android.core.exception.AuthenticationException
 import com.stripe.android.core.exception.InvalidRequestException
+import com.stripe.android.core.model.StripeFile
 import com.stripe.android.core.model.StripeModel
 import com.stripe.android.core.networking.ApiRequest
-import com.stripe.android.exception.AuthenticationException
 import com.stripe.android.model.PaymentIntent
 import com.stripe.android.model.SetupIntent
 import com.stripe.android.model.Source
-import com.stripe.android.model.StripeFile
 import com.stripe.android.model.StripeParamsModel
 import com.stripe.android.model.WeChatPayNextAction
 import com.stripe.android.networking.StripeApiRepository