+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/EmailSpec.kt
@@ -9,6 +9,6 @@ object EmailSpec : SectionFieldSpec(IdentifierSpec.Email) {
     fun transform(email: String?): SectionFieldElement =
         EmailElement(
             this.identifier,
-            TextFieldController(EmailConfig(), initialValue = email),
+            SimpleTextFieldController(EmailConfig(), initialValue = email),
         )
 }