+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/SectionUI.kt
@@ -66,11 +66,13 @@ fun SectionCard(
         border = PaymentsTheme.getBorderStroke(isSelected),
         // TODO(skyler-stripe): this will change when we add shadow configurations.
         elevation = if (isSelected) 1.5.dp else 0.dp,
-        backgroundColor = PaymentsTheme.colors.colorComponentBackground,
+        backgroundColor = PaymentsTheme.colors.component,
         shape = PaymentsTheme.shapes.material.medium,
         modifier = modifier
     ) {
-        content()
+        Column {
+            content()
+        }
     }
 }