+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/forms/TransformSpecToElements.kt
@@ -1,5 +1,6 @@
 package com.stripe.android.ui.core.forms
 
+import android.content.Context
 import com.stripe.android.ui.core.Amount
 import com.stripe.android.ui.core.address.AddressFieldElementRepository
 import com.stripe.android.ui.core.elements.AddressSpec
@@ -10,6 +11,8 @@ import com.stripe.android.ui.core.elements.AuBecsDebitMandateTextSpec
 import com.stripe.android.ui.core.elements.BankDropdownSpec
 import com.stripe.android.ui.core.elements.BankRepository
 import com.stripe.android.ui.core.elements.BsbSpec
+import com.stripe.android.ui.core.elements.CardBillingSpec
+import com.stripe.android.ui.core.elements.CardDetailsSpec
 import com.stripe.android.ui.core.elements.CountrySpec
 import com.stripe.android.ui.core.elements.EmailSpec
 import com.stripe.android.ui.core.elements.EmptyFormElement
@@ -41,7 +44,8 @@ class TransformSpecToElements(
     private val amount: Amount?,
     private val country: String?,
     private val saveForFutureUseInitialValue: Boolean,
-    private val merchantName: String
+    private val merchantName: String,
+    private val context: Context
 ) {
     fun transform(
         list: List<FormItemSpec>
@@ -125,6 +129,8 @@ class TransformSpecToElements(
                     currencyCode,
                     country
                 )
+                is CardDetailsSpec -> it.transform(context)
+                is CardBillingSpec -> it.transform(addressRepository)
                 is AuBankAccountNumberSpec -> it.transform()
             }
         }